spotter-sd/gemini_agent/pyproject.toml
[tool.poetry]
name = "gemini-agent"
version = "0.1.0"
description = ""
authors = ["Robert Goll <robertrover811@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uagents = "^0.11.1"
genai = "^2.1.0"
pillow = "^10.3.0"
fastapi = "^0.110.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
